+++ android/src/main/java/com/bitmovin/player/reactnative/analytics/conviva/BitmovinPlayerReactNativeAnalyticsConvivaPackage.kt
@@ -7,7 +7,10 @@ import com.facebook.react.uimanager.ViewManager
 
 class BitmovinPlayerReactNativeAnalyticsConvivaPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return listOf(BitmovinPlayerReactNativeAnalyticsConvivaModule(reactContext))
+        return listOf(
+            BitmovinPlayerReactNativeAnalyticsConvivaModule(reactContext),
+            BitmovinPlayerReactNativeAnalyticsConvivaSsaiModule(reactContext),
+        )
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {