+++ ios/BitmovinPlayerReactNativeAnalyticsConviva.swift
@@ -1,6 +1,5 @@
 import BitmovinConvivaAnalytics
 import BitmovinPlayerCore
-import ConvivaSDK
 import React
 import RNBitmovinPlayer
 
@@ -25,7 +24,7 @@ public class BitmovinPlayerReactNativeAnalyticsConviva: NSObject, RCTBridgeModul
         bridge.uiManager.methodQueue
     }
 
-    private func retrieve(_ nativeId: NativeId) -> ConvivaAnalytics? {
+    internal func retrieve(_ nativeId: NativeId) -> ConvivaAnalytics? {
         convivaAnalyticsInstances[nativeId]
     }