+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.1.0]
+
+### Added
+
+- `ConvivaAnalytics.ssai` namespace to enable server side ad tracking
+
+### Changed
+
+- Updated the `bitmovin-player-react-native` dependency to `0.27.1`.
+- Updated the Bitmovin Player Conviva Analytics Integration for Android dependency to `2.5.0`.
+- Updated the Bitmovin Player Conviva Analytics Integration for iOS dependency to `3.3.0`.
+
 ## [1.0.0] - 2024-06-14
 
 Adds support for Conviva Analytics with Bitmovin Player React Native SDK.